+++ packet/float_packet_field.h
@@ -9,9 +9,9 @@ public:
     explicit FloatPacketField(const QString &name,
                               const QString &type,
                               quint32 size,
-                              float min,
-                              float max,
-                              float initialValue = 0,
+                              double min,
+                              double max,
+                              double initialValue = 0,
                               quint8 decimals = 2,
                               FieldSource source = FieldSource::Constant,
                               QDataStream::ByteOrder byteOrder = QDataStream::LittleEndian);
@@ -23,19 +23,19 @@ public:
     static std::unique_ptr<FloatPacketField> fromJson(const QJsonObject &obj,
                                                       QString *outError = nullptr);
 
-    float min() const;
-    float max() const;
+    double min() const;
+    double max() const;
     quint8 decimals() const;
 
-    float value() const;
-    void setValue(float newValue);
+    double value() const;
+    void setValue(double newValue);
 
 private:
-    float m_max; //  maximum value
-    float m_min; // minimum value
+    double m_max; //  maximum value
+    double m_min; // minimum value
     quint8 m_decimals = 2;
 
-    float m_value = 0;
+    double m_value = 0;
 };
 
 #endif // FLOAT_PACKET_FIELD_H
